+++ prospector/tools/pep257/__init__.py
@@ -5,13 +5,13 @@ from __future__ import absolute_import
 # pep257 version 0.4.1 sets global log level to debug
 # which causes it to spaff the output with tokenizing
 # information.
-import pep257
-if hasattr(pep257, 'log'):
+import pydocstyle  # noqa
+if hasattr(pydocstyle, 'log'):  # noqa
     def dummy_log(*args, **kwargs):  # noqa
         pass
-    pep257.log.debug = dummy_log
+    pydocstyle.log.debug = dummy_log
 
-from pep257 import PEP257Checker, AllError
+from pydocstyle import PEP257Checker, AllError
 from prospector.message import Location, Message
 from prospector.tools.base import ToolBase